flask/requirements.txt
adal==1.2.4
asgiref==3.2.3
beautifulsoup4==4.9.1
bs4==0.0.1
CacheControl==0.12.6
cached-property==1.5.1
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
crypto==1.4.1
cryptography
defusedxml==0.6.0
dialogflow==1.0.0
dialogflow-fulfillment==0.3.0
dnspython==1.16.0
exchangelib==3.2.0
firebase==3.0.1
firebase-admin==4.3.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
gcloud==0.18.3
google==3.0.0
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==19.7.1
httplib2==0.18.1
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jwcrypto==0.7
line-bot-sdk==1.16.0
lxml==4.5.1
MarkupSafe==1.1.1
msgpack==1.0.0
mysqlclient==1.4.6
Naked==0.1.31
ndg-httpsclient==0.5.1
ntlm-auth==1.5.0
oauth2client==4.1.3
oauthlib==3.1.0
pip-autoremove==0.9.1
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-firebase
python-jwt==3.2.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
selenium==3.141.0
shellescape==3.8.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
sqlparse==0.3.0
sseclient==0.0.26
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
